+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this library are documented in this file.
 
 ### Bug Fixes
 
+- Account for `MapPlot` custom domain that crosses anti-meridian (#655).
 - Reduce needless lat/lon precision with Jabber messages (#656).
 
 ## **1.14.0** (15 Sep 2022)